restructure models and schemas into separate modules
