+++ api/models/file.py
@@ -1,0 +1,17 @@
+from datetime import datetime
+from sqlalchemy import Column, String, JSON, DateTime
+
+from db.database import Base
+
+
+class FileMetadata(Base):
+    __tablename__ = "file_metadata"
+
+    id = Column(String, primary_key=True)
+    filename = Column(String)
+    content_type = Column(String)
+    tika_metadata = Column(JSON)
+    content = Column(String)
+    user_id = Column(String)
+    created_at = Column(DateTime, default=datetime.utcnow)
+    updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
